Fixed the tests
Run `gradle check` to run. No IDE support just yet.

+++ scdl/src/main/java/net/rdrei/android/scdl2/api/entity/UserEntity.java
@@ -48,7 +48,7 @@ public class UserEntity implements SoundcloudEntity, Parcelable {
 	}
 
 	private UserEntity(final Parcel in) {
-		id = in.readInt();
+		id = in.readLong();
 		username = in.readString();
 		uri = in.readString();
 	}
